+++ src/sip_layer.c
@@ -30,7 +30,7 @@ static char const ident[]="$Id$";
  * argument for a number of functions used by siproxd.
  */
 
-int sip_message_parse(osip_message_t * sip, const char *buf, int len) {
+int sip_message_parse(osip_message_t * sip, const char *buf, size_t len) {
 #ifdef HAVE_FUNC_OSIP_MESSAGE_PARSE_3
    return osip_message_parse(sip, buf, len);
 #else 
@@ -38,7 +38,7 @@ int sip_message_parse(osip_message_t * sip, const char *buf, int len) {
 #endif
 }
 
-int sip_message_to_str(osip_message_t * sip, char **dest, int *len) {
+int sip_message_to_str(osip_message_t * sip, char **dest, size_t *len) {
 #ifdef  HAVE_FUNC_OSIP_MESSAGE_TO_STR_3
    int sts;
    sts = osip_message_to_str(sip, dest, len);
@@ -56,7 +56,7 @@ int sip_message_to_str(osip_message_t * sip, char **dest, int *len) {
 #endif
 }
 
-int sip_body_to_str(const osip_body_t * body, char **dest, int *len) {
+int sip_body_to_str(const osip_body_t * body, char **dest, size_t *len) {
 #ifdef  HAVE_FUNC_OSIP_BODY_TO_STR_3
    int sts;
    sts = osip_body_to_str(body, dest, len);
@@ -74,7 +74,7 @@ int sip_body_to_str(const osip_body_t * body, char **dest, int *len) {
 #endif
 }
 
-int sip_message_set_body(osip_message_t * sip, const char *buf, int len) {
+int sip_message_set_body(osip_message_t * sip, const char *buf, size_t len) {
 #ifdef  HAVE_FUNC_OSIP_MESSAGE_SET_BODY_3
    return osip_message_set_body(sip, buf, len);
 #else
